+++ scripts/dialogs/themer.mjs
@@ -72,7 +72,7 @@ export class ThemeDialog extends FormApplication {
 			template: `./modules/${MODULE.ID}/templates/themer.hbs`,
 			width: $(window).width() > 960 ? 960 : $(window).width() - 100,
 			height: $(window).height() > 800 ? 800 : $(window).height() - 100
-		});
+		}, {});
 	}
   
 	getData() {